feat: use DjangoFilterAndOffsetConnectionField with messages query
KK-1212.

Refactor the DjangoFilterAndOffsetConnectionField to be a common filter.
Add the DjangoFilterAndOffsetConnectionField for messaging messages
query.
